@@ -17,15 +17,14 @@
     "@plone/providers": "workspace:*",
     "@plone/registry": "workspace:*",
     "@plone/blocks": "workspace:*",
-    "@tanstack/react-query": "^5.37.1",
-    "@tanstack/react-router": "^1.28.2",
-    "@tanstack/react-router-server": "^1.28.2",
-    "@tanstack/router-devtools": "^1.28.2",
-    "@tanstack/router-vite-plugin": "^1.28.2",
+    "@tanstack/react-query": "^5.59.0",
+    "@tanstack/react-router": "^1.34.9",
+    "@tanstack/start": "^1.34.9",
+    "@tanstack/router-vite-plugin": "^1.34.8",
     "axios": "^1.6.5",
     "get-port": "^7.0.0",
     "react": "^18.2.0",
-    "react-aria-components": "^1.1.1",
+    "react-aria-components": "^1.4.0",
     "react-dom": "^18.2.0",
     "sirv": "^2.0.4"
   },
@@ -34,7 +33,8 @@
     "@babel/generator": "^7.23.6",
     "@plone/types": "workspace:*",
     "@rollup/plugin-babel": "^6.0.4",
-    "@tanstack/react-query-devtools": "^5.37.1",
+    "@tanstack/react-query-devtools": "^5.59.0",
+    "@tanstack/router-devtools": "^1.34.9",
     "@types/express": "^4.17.21",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
@@ -44,7 +44,7 @@
     "isbot": "^4.3.0",
     "node-fetch": "^3.3.2",
     "serve-static": "^1.15.0",
-    "typescript": "^5.4.2",
+    "typescript": "^5.6.3",
     "vite": "^5.2.9",
     "vite-plugin-babel": "^1.2.0"
   }
