@@ -8,7 +8,7 @@
     }
   ],
   "license": "MIT",
-  "version": "2.0.0-alpha.12",
+  "version": "2.0.0-alpha.16",
   "repository": {
     "type": "git",
     "url": "http://github.com/plone/components.git"
@@ -110,9 +110,9 @@
     "parcel": "^2.12.0",
     "release-it": "17.1.1",
     "storybook": "^8.0.4",
-    "typescript": "^5.4.5",
+    "typescript": "^5.6.3",
     "vite": "^5.4.8",
-    "vitest": "^2.1.2",
+    "vitest": "^2.1.3",
     "vitest-axe": "^0.1.0"
   },
   "dependencies": {
