@@ -2,21 +2,32 @@ import { cssBundleHref } from '@remix-run/css-bundle';
 import type { LinksFunction } from '@remix-run/node';
 import {
   Links,
-  LiveReload,
   Meta,
   Outlet,
   Scripts,
   ScrollRestoration,
+  useHref,
+  useLocation,
+  useNavigate,
+  useParams,
 } from '@remix-run/react';
 import { useState } from 'react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { PloneClientProvider } from '@plone/providers';
+import { QueryClient } from '@tanstack/react-query';
 import PloneClient from '@plone/client';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+
+import '@plone/components/dist/basic.css';
+import { flattenToAppURL } from './utils';
+import { PloneProvider } from '@plone/providers';
+import config from '@plone/registry';
 
 export const links: LinksFunction = () => [
   ...(cssBundleHref ? [{ rel: 'stylesheet', href: cssBundleHref }] : []),
 ];
+
+function useHrefLocal(to: string) {
+  return useHref(flattenToAppURL(to));
+}
 
 export default function App() {
   const [queryClient] = useState(
@@ -34,9 +45,15 @@ export default function App() {
 
   const [ploneClient] = useState(() =>
     PloneClient.initialize({
-      apiPath: 'http://localhost:8080/Plone',
+      apiPath: config.settings.apiPath,
     }),
   );
+
+  const RRNavigate = useNavigate();
+
+  const navigate = (to: string) => {
+    return RRNavigate(flattenToAppURL(to));
+  };
 
   return (
     <html lang="en">
@@ -47,15 +64,19 @@ export default function App() {
         <Links />
       </head>
       <body>
-        <PloneClientProvider client={ploneClient}>
-          <QueryClientProvider client={queryClient}>
-            <Outlet />
-            <ScrollRestoration />
-            <Scripts />
-            <LiveReload />
-            <ReactQueryDevtools initialIsOpen={false} />
-          </QueryClientProvider>
-        </PloneClientProvider>
+        <PloneProvider
+          ploneClient={ploneClient}
+          queryClient={queryClient}
+          useLocation={useLocation}
+          useParams={useParams}
+          useHref={useHrefLocal}
+          navigate={navigate}
+        >
+          <Outlet />
+          <ScrollRestoration />
+          <Scripts />
+          <ReactQueryDevtools initialIsOpen={false} />
+        </PloneProvider>
       </body>
     </html>
   );
