@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
 
-import { StartClient } from '@tanstack/react-router-server/client';
+import { StartClient } from '@tanstack/start/client';
 import { createRouter } from './router';
 
 import './config';
