@@ -3,7 +3,7 @@ import ReactDOMServer from 'react-dom/server';
 import { createMemoryHistory } from '@tanstack/react-router';
 import { ServerResponse } from 'http';
 import express from 'express';
-import { StartServer } from '@tanstack/react-router-server/server';
+import { StartServer } from '@tanstack/start/server';
 import { createRouter } from './router';
 import './config';
 
