@@ -37,14 +37,23 @@ const IndexRoute = IndexImport.update({
 declare module '@tanstack/react-router' {
   interface FileRoutesByPath {
     '/': {
+      id: '/'
+      path: '/'
+      fullPath: '/'
       preLoaderRoute: typeof IndexImport
       parentRoute: typeof rootRoute
     }
     '/$': {
+      id: '/$'
+      path: '/$'
+      fullPath: '/$'
       preLoaderRoute: typeof SplatImport
       parentRoute: typeof rootRoute
     }
     '/error': {
+      id: '/error'
+      path: '/error'
+      fullPath: '/error'
       preLoaderRoute: typeof ErrorImport
       parentRoute: typeof rootRoute
     }
@@ -53,10 +62,34 @@ declare module '@tanstack/react-router' {
 
 // Create and export the route tree
 
-export const routeTree = rootRoute.addChildren([
+export const routeTree = rootRoute.addChildren({
   IndexRoute,
   SplatRoute,
   ErrorRoute,
-])
+})
 
 /* prettier-ignore-end */
+
+/* ROUTE_MANIFEST_START
+{
+  "routes": {
+    "__root__": {
+      "filePath": "__root.tsx",
+      "children": [
+        "/",
+        "/$",
+        "/error"
+      ]
+    },
+    "/": {
+      "filePath": "index.tsx"
+    },
+    "/$": {
+      "filePath": "$.tsx"
+    },
+    "/error": {
+      "filePath": "error.tsx"
+    }
+  }
+}
+ROUTE_MANIFEST_END */
