@@ -1,16 +1,9 @@
-import {
-  QueryClient,
-  QueryClientProvider,
-  dehydrate,
-  hydrate,
-} from '@tanstack/react-query';
+import { QueryClient, dehydrate, hydrate } from '@tanstack/react-query';
 import PloneClient from '@plone/client';
 import { createRouter as createReactRouter } from '@tanstack/react-router';
 
 import { routeTree } from './routeTree.gen';
 import { PloneClientProvider } from '@plone/providers';
-import { FlattenToAppURLProvider } from '@plone/components';
-import { flattenToAppURL } from './utils';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -45,12 +38,8 @@ export function createRouter() {
     },
     Wrap: ({ children }: { children: any }) => {
       return (
-        <PloneClientProvider client={ploneClient}>
-          <QueryClientProvider client={queryClient}>
-            <FlattenToAppURLProvider flattenToAppURL={flattenToAppURL}>
-              {children}
-            </FlattenToAppURLProvider>
-          </QueryClientProvider>
+        <PloneClientProvider client={ploneClient} queryClient={queryClient}>
+          {children}
         </PloneClientProvider>
       );
     },
