@@ -1,5 +1,5 @@
-import { createFileRoute } from '@tanstack/react-router';
 import * as React from 'react';
+import { createFileRoute } from '@tanstack/react-router';
 import { flattenToAppURL } from '../utils';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { usePloneClient } from '@plone/providers';
@@ -23,6 +23,7 @@ function IndexComponent() {
   const { data } = useSuspenseQuery(
     getContentQuery({ path: flattenToAppURL('/'), expand }),
   );
+
   return (
     <>
       <Breadcrumbs
