@@ -1,6 +1,6 @@
-export * from './RouterLocation';
+export * from './AppRouter';
 export * from './PloneClient';
-// FlattenToAppURL provider should live on the components
-// but @plone/components can't depend on other packages
-// we are proxying it here for convenience and centralization
-export { FlattenToAppURLProvider, useFlattenToAppURL } from '@plone/components';
+export * from './PloneProvider';
+
+// Proxying RouterProvider from react-aria-components
+export { RouterProvider } from 'react-aria-components';
